backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "appprecos-backend"
version = "2.0.0"
description = "AppPrecos Backend API - NFCe Data Extraction with Supabase PostgreSQL"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "AppPrecos Team", email = "contact@appprecos.com" },
]
license = { text = "MIT" }
keywords = ["nfce", "supabase", "flask", "postgresql", "api"]

dependencies = [
    "Flask==3.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "playwright==1.55.0",
    "openpyxl==3.1.2",
    "requests==2.32.3",
    "supabase==2.3.5",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/appprecos"
Documentation = "https://github.com/yourusername/appprecos/wiki"
Repository = "https://github.com/yourusername/appprecos"

[tool.setuptools]
packages = ["appprecos"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=backend"
